Ran into another problem where helm wouldn't work because it couldn't find the tiller pod.
master aka apiserver can't find pods issued on other nodes. So
`kubectl describe pod XXX` will not work if the pod is on another node,
but will work if the pod is running on the master node.

It was because the internal nodeip (shown when you run
`kubectl get nodes -o wide` was wrong because it was using the default
interface, but we needed it to use the bridged interface.

Fixed the problem by changing the kubelet environment file at
`/etc/default/kubelet`, putting `KUBELET_EXTRA_ARGS=--node-ip=192.168.99.2x`
so that it knows what its IP should be. Configured this with ansible.